requirement_finetuning.txt
fire=0.7.0
simple-parsing=0.1.7
pyyaml=6.0.2
mistral-common>=1.3.1
safetensors=0.5.3
tensorboard
tqdm=4.67.1
unsloth=2025.5.6
unsloth_zoo=2025.5.7
huggingface-hub=0.31.4
atasets=3.6.0
trl=0.15.2


 numpy==1.26.4
 torch==2.7.0 
 torchvision==0.22.0
 transformers==4.51.3
